Add frame icon for take screenshot
