+++ extensions/deviceicon/Makefile.am
@@ -3,6 +3,7 @@ sugardir = $(pkgdatadir)/extensions/deviceicon
 sugar_PYTHON = 		\
 	__init__.py	\
 	battery.py	\
+	display.py	\
 	frame.py	\
 	network.py	\
 	speaker.py	\
